02/06/2024

+++ public/assets/js/modul_ac/chart-ac-cari.js
@@ -4,7 +4,6 @@ $(document).ready(function () {
 
         var tahun = $("#updateTahun").val();
         const hapusSemua = $("#updateTahun").data("hapussemua");
-        const hapusChart = $("#updateTahun").data("urlhapuschartajax");
         const urlCariTahun = $("#updateTahun").data("urlcaricart");
 
         $.ajax({
@@ -17,18 +16,23 @@ $(document).ready(function () {
                 let card = "";
                 data.map((item) => {
                     $("#modalChartCari").modal("show");
-                    card += updateChartAc(item, hapusChart);
+                    card += updateChartAc(item);
+
                     $("#headerBtnDelete").empty();
-                    //   $("#headerBtnDelete").append(
-                    //       `<a href='${hapusSemua}' class='btn btn-danger mb-3'>Hapus data ${item.tahun}</a>`
-                    //   );
-                    $("#headerBtnDelete").append(
-                        `<form action="${hapusSemua}" method="get">
-                                <input hidden value="${item.tahun}" name="deleteAllChart">
-                                    <button type="submit"
-                                        class="btn btn-danger mb-3">Hapus data ${item.tahun}</button>
-                            </form>`
-                    );
+                    const superAdmin =
+                        $("#superAdminStatus").data("super-admin");
+                    if (superAdmin) {
+                        $("#headerBtnDelete").append(
+                            `<form id="formDelete" action="${hapusSemua}" method="get">
+                            <input hidden value="${item.tahun}" name="deleteAllChart">
+                            <button type="submit" class="btn btn-danger mb-3" id="btnChartAllDelete">Hapus data ${item.tahun}</button>
+                        </form>`
+                        );
+                    } else {
+                        $("#titleCariChartTahun").text(
+                            `Data tahun ${item.tahun}`
+                        );
+                    }
                 });
                 $("#modalBodyChartCari").html(card);
             },
@@ -39,20 +43,28 @@ $(document).ready(function () {
     });
 });
 
-function updateChartAc(e, hapus) {
-    const urlChart = hapus + "/" + e.id + "/" + e.tahun;
-
-    return `<tr>
-                  
-                  <td>${e.tahun}</td>
-                  <td>${e.bulan}</td>
-                  <td>${e.total}</td>
-                  <td>
-
-                        <a href="${urlChart}"
-                                            id="btnDeleteChartTb">
-                                            <i class='bx bx-trash fs-6 text-danger'></i>
-                                        </a>
-                  </td>
+function updateChartAc(val) {
+    return `<tr>                  
+                  <td>${val.bulan}</td>
+                  <td>${val.total}</td>                  
               </tr>`;
 }
+
+$(document).on("submit", "#formDelete", function (e) {
+    e.preventDefault();
+    const tahunToDelete = $("input[name='deleteAllChart']").val();
+    const href = $(this).attr("action");
+    Swal.fire({
+        title: "Apa kamu yakin?",
+        text: `Ingin menghapus data tahun ${tahunToDelete}?`,
+        type: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Hapus!",
+    }).then((result) => {
+        if (result.isConfirmed) {
+            window.location.href = href + "?" + $(this).serialize();
+        }
+    });
+});
